payroll_period: work-arround for payslip.details_by_salary_rule_category not being computed
In this module we do some extra work in compute_sheet() for which we
need the details_by_salary_rule_category field to be computed. But
it seems that at the time of compute_sheet() this field is not computed.
So we roll our own method getting this same information: get_categories_dict()