Massive module rename
Most of my HR modules have the same name as in 7.0. This creates a
module name conflic in the Odoo App Store because those modules
are under my name and these modules are under TREVI's account. Therefore,
the new modules do not appear in the app store. Most of the changes are
straight renames with one notable exception. I have folded hr_security
into hr_simplify and renamed it to trevi_hr_usability.