payroll_period: don't require pay period schedule on hr.contract
This was causing the tests on other unrelated module tests to fail
because they were trying to create contracts in their tests and
the pps_id field was raising NULL VALUE errors. The obvious thing
to do then is the drop the required attribute in the model and
instead enforce it in the form view instead.