[IMP] resource_schedule - Add the capability of flex shifts
A Flex Shift is, as it's name implies, a more flexible type of shift.
In a flex shift you can define the core period the employee needs
to be at work, but leave the choice of when to come in and when
to clock out at the employees discretion. You can futher restrict
the shift so that the employee must work a certain number of
hours per day, per week, or a daily average based on the weekly
total. You can also make it less restrictive by ommitting the
core period alltogether and simply specifying only a mandatory
scheduled hours/day or per week (or both).