[ADD] payroll_default_salary_rules: basic set of necessary salary rules and categories to go with them
This module provides a very small set of essential salary rules and categories.
It also allows the administrator to set the specific salary rule and
salary rule category codes to be used in configuration settings. This
should make developing payroll modules and reports easier since they
can rely on the settings instead of specific codes.